+++ test/Record.ts
@@ -13,6 +13,10 @@ import { showString } from '../src/Show'
 const p = (n: number) => n > 2
 
 describe('Record', () => {
+  it('collect', () => {
+    assert.deepStrictEqual(R.collect((_, a) => a)({ a: 1, b: 2 }), [1, 2])
+  })
+
   it('getMonoid', () => {
     const d1 = { k1: 1, k2: 3 }
     const d2 = { k2: 2, k3: 4 }
@@ -23,7 +27,10 @@ describe('Record', () => {
   it('map', () => {
     const d1 = { k1: 1, k2: 2 }
     const double = (n: number): number => n * 2
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.map(d1, double), { k1: 2, k2: 4 })
+
+    assert.deepStrictEqual(R.map(double)(d1), { k1: 2, k2: 4 })
   })
 
   it('reduce', () => {
@@ -36,21 +43,34 @@ describe('Record', () => {
   it('foldMap', () => {
     const foldMap = R.foldMap(monoidString)
     const x1 = { a: 'a', b: 'b' }
-    const f1 = identity
+    const f1 = (s: string): string => s
+    // tslint:disable-next-line: deprecation
     assert.strictEqual(foldMap(x1, f1), 'ab')
+
+    assert.strictEqual(foldMap(f1)(x1), 'ab')
   })
 
-  it('foldr', () => {
-    const foldr = R.foldr
+  it('foldr / reduceRight', () => {
     const x1 = { a: 'a', b: 'b' }
     const init1 = ''
     const f1 = (a: string, acc: string) => acc + a
-    assert.strictEqual(foldr(x1, init1, f1), 'ba')
+    // tslint:disable-next-line: deprecation
+    assert.strictEqual(R.foldr(x1, init1, f1), 'ba')
+
+    assert.strictEqual(R.reduceRight(init1, f1)(x1), 'ba')
   })
 
   it('traverse', () => {
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.traverse(option)({ k1: 1, k2: 2 }, n => (n <= 2 ? some(n) : none)), some({ k1: 1, k2: 2 }))
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.traverse(option)({ k1: 1, k2: 2 }, n => (n >= 2 ? some(n) : none)), none)
+
+    assert.deepStrictEqual(
+      R.traverse2v(option)((n: number) => (n <= 2 ? some(n) : none))({ k1: 1, k2: 2 }),
+      some({ k1: 1, k2: 2 })
+    )
+    assert.deepStrictEqual(R.traverse2v(option)((n: number) => (n >= 2 ? some(n) : none))({ k1: 1, k2: 2 }), none)
   })
 
   it('sequence', () => {
@@ -92,16 +112,20 @@ describe('Record', () => {
   })
 
   it('mapWithKey', () => {
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.mapWithKey({ aa: 1 }, (k, a) => a + k.length), { aa: 3 })
   })
 
   it('traverseWithKey', () => {
     const d1 = { k1: 1, k2: 2 }
+    // tslint:disable-next-line: deprecation
     const t1 = R.traverseWithKey(option)(d1, (k, n): Option<number> => (k !== 'k1' ? some(n) : none))
     assert.deepStrictEqual(t1, none)
     const d2 = { k1: 2, k2: 3 }
+    // tslint:disable-next-line: deprecation
     const t2 = R.traverseWithKey(option)(d2, (k, n): Option<number> => (k !== 'k3' ? some(n) : none))
     assert.deepStrictEqual(t2, some({ k1: 2, k2: 3 }))
+    // tslint:disable-next-line: deprecation
     const t3 = R.traverseWithKey(option)({}, (k, n): Option<number> => none)
     assert.strictEqual(t3.getOrElse({}), R.empty)
   })
@@ -116,24 +140,29 @@ describe('Record', () => {
     assert.strictEqual(R.isEmpty({ a: 1 }), false)
   })
 
-  it('insert', () => {
-    assert.deepStrictEqual(R.insert('a', 1, {}), { a: 1 })
-    assert.deepStrictEqual(R.insert('c', 3, { a: 1, b: 2 }), { a: 1, b: 2, c: 3 })
+  it('insertAt', () => {
+    assert.deepStrictEqual(R.insertAt('a', 1)({}), { a: 1 })
+    assert.deepStrictEqual(R.insertAt('c', 3)({ a: 1, b: 2 }), { a: 1, b: 2, c: 3 })
     // should return the same reference if the value is already there
     const x = { a: 1 }
-    assert.strictEqual(R.insert('a', 1, x), x)
+    assert.strictEqual(R.insertAt('a', 1)(x), x)
   })
 
-  it('remove', () => {
-    assert.deepStrictEqual(R.remove('a', { a: 1, b: 2 }), { b: 2 })
+  it('deleteAt', () => {
+    assert.deepStrictEqual(R.deleteAt('a')({ a: 1, b: 2 }), { b: 2 })
     // should return the same reference if the key is missing
     const x = { a: 1 }
-    assert.strictEqual(R.remove('b', x), x)
+    assert.strictEqual(R.deleteAt('b')(x), x)
   })
 
   it('pop', () => {
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.pop('a', { a: 1, b: 2 }), some([1, { b: 2 }]))
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.pop('c', { a: 1, b: 2 }), none)
+
+    assert.deepStrictEqual(R.pop('a')({ a: 1, b: 2 }), some([1, { b: 2 }]))
+    assert.deepStrictEqual(R.pop('c')({ a: 1, b: 2 }), none)
   })
 
   it('compact', () => {
@@ -149,29 +178,33 @@ describe('Record', () => {
 
   it('filter', () => {
     const d = { a: 1, b: 3 }
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.filter(d, p), { b: 3 })
-
-    // refinements
-    const isNumber = (u: string | number): u is number => typeof u === 'number'
-    const y: Record<string, string | number> = { a: 1, b: 'foo' }
-    const actual = R.filter(y, isNumber)
-    assert.deepStrictEqual(actual, { a: 1 })
-
-    assert.strictEqual(R.filter(y, _ => true), y)
-
-    const x = Object.assign(Object.create({ c: true }), { a: 1, b: 'foo' })
-    assert.deepStrictEqual(R.filter(x, isNumber), { a: 1 })
+    assert.deepStrictEqual(R.filter(p)(d), { b: 3 })
   })
 
   it('filterMap', () => {
     const f = (n: number) => (p(n) ? some(n + 1) : none)
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.filterMap({}, f), {})
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.filterMap({ a: 1, b: 3 }, f), { b: 4 })
+
+    assert.deepStrictEqual(R.filterMap(f)({}), {})
+    assert.deepStrictEqual(R.filterMap(f)({ a: 1, b: 3 }), { b: 4 })
   })
 
   it('partition', () => {
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partition({}, p), { left: {}, right: {} })
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partition({ a: 1, b: 3 }, p), {
+      left: { a: 1 },
+      right: { b: 3 }
+    })
+
+    assert.deepStrictEqual(R.partition(p)({}), { left: {}, right: {} })
+    assert.deepStrictEqual(R.partition(p)({ a: 1, b: 3 }), {
       left: { a: 1 },
       right: { b: 3 }
     })
@@ -179,8 +212,16 @@ describe('Record', () => {
 
   it('partitionMap', () => {
     const f = (n: number) => (p(n) ? right(n + 1) : left(n - 1))
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partitionMap({}, f), { left: {}, right: {} })
+    // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partitionMap({ a: 1, b: 3 }, f), {
+      left: { a: 0 },
+      right: { b: 4 }
+    })
+
+    assert.deepStrictEqual(R.partitionMap(f)({}), { left: {}, right: {} })
+    assert.deepStrictEqual(R.partitionMap(f)({ a: 1, b: 3 }), {
       left: { a: 0 },
       right: { b: 4 }
     })
@@ -200,21 +241,30 @@ describe('Record', () => {
     assert.deepStrictEqual(wiltIdentity({ a: 1, b: 3 }, f), new I.Identity({ left: { a: 0 }, right: { b: 4 } }))
   })
 
-  it('reduceWithKey', () => {
-    const d1 = { k1: 'a', k2: 'b' }
-    assert.strictEqual(R.reduceWithKey(d1, '', (k, b, a) => b + k + a), 'k1ak2b')
-    const d2 = { k2: 'b', k1: 'a' }
-    assert.strictEqual(R.reduceWithKey(d2, '', (k, b, a) => b + k + a), 'k1ak2b')
+  it('reduceWithKey / reduceWithIndex', () => {
+    // tslint:disable-next-line: deprecation
+    assert.strictEqual(R.reduceWithKey({ k1: 'a', k2: 'b' }, '', (k, b, a) => b + k + a), 'k1ak2b')
+    // tslint:disable-next-line: deprecation
+    assert.strictEqual(R.reduceWithKey({ k2: 'b', k1: 'a' }, '', (k, b, a) => b + k + a), 'k1ak2b')
+
+    assert.strictEqual(R.reduceWithIndex('', (k, b, a) => b + k + a)({ k1: 'a', k2: 'b' }), 'k1ak2b')
+    assert.strictEqual(R.reduceWithIndex('', (k, b, a) => b + k + a)({ k2: 'b', k1: 'a' }), 'k1ak2b')
   })
 
-  it('foldMapWithKey', () => {
+  it('foldMapWithKey / foldMapWithIndex', () => {
     const x1 = { k1: 'a', k2: 'b' }
+    // tslint:disable-next-line: deprecation
     assert.strictEqual(R.foldMapWithKey(monoidString)(x1, (k, a) => k + a), 'k1ak2b')
+
+    assert.strictEqual(R.foldMapWithIndex(monoidString)((k, a) => k + a)(x1), 'k1ak2b')
   })
 
-  it('foldrWithKey', () => {
+  it('foldrWithKey / reduceRightWithIndex', () => {
     const x1 = { k1: 'a', k2: 'b' }
+    // tslint:disable-next-line: deprecation
     assert.strictEqual(R.foldrWithKey(x1, '', (k, a, b) => b + k + a), 'k2bk1a')
+
+    assert.strictEqual(R.reduceRightWithIndex('', (k, a, b) => b + k + a)(x1), 'k2bk1a')
   })
 
   it('every', () => {
@@ -245,17 +295,21 @@ describe('Record', () => {
     const f = (k: string, a: string) => (k === 'a' ? right(a.length) : left(a))
     // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partitionMapWithIndex(x, f), expected)
+
+    assert.deepStrictEqual(R.partitionMapWithIndex(f)(x), expected)
   })
 
   it('partitionWithIndex', () => {
     const x = { a: 'a', b: 'b' }
-    const f = (k: string) => k === 'a'
+    const f = (k: string, _: string) => k === 'a'
     const expected = {
       left: { b: 'b' },
       right: { a: 'a' }
     }
     // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.partitionWithIndex(x, f), expected)
+
+    assert.deepStrictEqual(R.partitionWithIndex(f)(x), expected)
   })
 
   it('filterMapWithIndex', () => {
@@ -266,6 +320,8 @@ describe('Record', () => {
     }
     // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.filterMapWithIndex(x, f), expected)
+
+    assert.deepStrictEqual(R.filterMapWithIndex(f)(x), expected)
   })
 
   it('filterWithIndex', () => {
@@ -274,6 +330,12 @@ describe('Record', () => {
     const expected = { a: 'a' }
     // tslint:disable-next-line: deprecation
     assert.deepStrictEqual(R.filterWithIndex(x, f), expected)
+
+    assert.deepStrictEqual(R.filterWithIndex(f)(x), expected)
+    const x2 = Object.create({ c: 1 })
+    Object.assign(x2, x)
+    assert.deepStrictEqual(R.filterWithIndex(f)(x2), expected)
+    assert.strictEqual(R.filterWithIndex((k: string) => k === 'a' || k === 'b')(x), x)
   })
 
   it('fromFoldableMap', () => {
@@ -304,5 +366,26 @@ describe('Record', () => {
     assert.strictEqual(S.show({}), `{}`)
     assert.strictEqual(S.show({ a: 'a' }), `{ "a": "a" }`)
     assert.strictEqual(S.show({ a: 'a', b: 'b' }), `{ "a": "a", "b": "b" }`)
+  })
+
+  it('reduceRightWithIndex', () => {
+    const x1 = { k1: 'a', k2: 'b' }
+    assert.strictEqual(R.reduceRightWithIndex('', (k, a, b) => b + k + a)(x1), 'k2bk1a')
+  })
+
+  it('reduceWithIndex', () => {
+    const d1 = { k1: 'a', k2: 'b' }
+    assert.strictEqual(R.reduceWithIndex('', (k, b, a) => b + k + a)(d1), 'k1ak2b')
+    const d2 = { k2: 'b', k1: 'a' }
+    assert.strictEqual(R.reduceWithIndex('', (k, b, a) => b + k + a)(d2), 'k1ak2b')
+  })
+
+  it('foldMapWithIndex', () => {
+    const x1 = { k1: 'a', k2: 'b' }
+    assert.strictEqual(R.foldMapWithIndex(monoidString)((k, a) => k + a)(x1), 'k1ak2b')
+  })
+
+  it('mapWithIndex', () => {
+    assert.deepStrictEqual(R.mapWithIndex((k, a: number) => a + k.length)({ aa: 1 }), { aa: 3 })
   })
 })